% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract_two_samp_test.R
\name{abstract_two_sample_test}
\alias{abstract_two_sample_test}
\title{Abstract two sample test}
\usage{
abstract_two_sample_test(
  response_odm,
  grna_odm,
  response_grna_group_pairs,
  two_sample_test,
  progress,
  control_group = "ntc",
  cbind_res = FALSE
)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{grna_odm}{an ODM of either grna expressions (counts) or grna
assignments (logical)}

\item{response_grna_group_pairs}{a data frame with columns \code{response_id} and
\code{grna_group} giving the response ID / grna group pairs to analyze.}

\item{two_sample_test}{a two-sample test; should take as arguments (i) vector of expressions of target cells, (ii) vector of expressions of control cells, (iii) the indices of cells receiving the targeting grna, and (iv) the indices of the cells receiving the NT grnas.}

\item{progress}{print progress messages?}

\item{control_group}{the control group to use: "ntc" for cells containing an NTC or "compliment" for the compliment of the treatment cells}

\item{cbind_res}{cbind the result to \code{response_grna_group_pairs} (TRUE) or simply return the \code{response_grna_group_pairs} data frame with the vector of p-values appended (FALSE)?}
}
\value{
a data frame with columns \code{response_id}, \code{grna_group}, and \code{p_value}.
}
\description{
An abstract a two-sample test. Pass function \code{two_sample_test} to carry out a given two-sample test (e.g., a t-test or a permutation test).
}
\details{
The \code{grna_group} column of \code{response_grna_group_pairs} contains the names of grna groups.
These grna groups are assumed to be present in the \code{target} column of the feature
covariate matrix of \code{grna_odm}. This column should contain entries "non-targeting"
indicating the non-targeting grnas. If \code{target_type} is present as a column of the
feature covariate matrix of \code{grna_odm}, \code{target_type} is ignored.
}
\examples{
\dontrun{
two_sample_test <- function(target_cells, control_cells, response_id,
target_cell_indices, control_cell_indices) {
t.test(target_cells, control_cells)$p.value
}
response_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/gene")
grna_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/grna_assignment")
response_grna_group_pairs <-
 expand.grid(grna_group = c("CCNE2", "HS2"),
             response_id = sample(ondisc::get_feature_ids(response_odm), 50))
abstract_two_sample_test(response_odm, grna_odm, response_grna_group_pairs, two_sample_test, TRUE)
}
}
