% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weissman.R
\name{add_ENSG}
\alias{add_ENSG}
\title{Add ENSG prefix to a list of feature ids if necessary}
\usage{
add_ENSG(feature_ids)
}
\arguments{
\item{feature_id}{Vector of feature ids}
}
\value{
The same set of feature ids with ENSG prefixes added if necessary
}
\description{
This is somewhat of a hack needed for compatibility with perturbseq code.
}
