% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schraivogel_helpers.R
\name{de_MAST}
\alias{de_MAST}
\title{MAST DE function}
\usage{
de_MAST(pert_object)
}
\arguments{
\item{pert_object}{SingleCellExperiment object containing gene expression data and cell groupings
in colData. The perturbation to be tested is assumed to be the first column in colData!}
}
\description{
Perform differential expression analysis between perturbed and control cells using MAST.
}
\details{
The first column in colData is expected to provide the perturbation status as a factor with two
levels, where the first level is non-perturbed cells. If this is not the case, the function will
attempt to convert the first column to a two level factor, with any risks that entails.
Any other columns in colData will be used as covariates during modelling, but only the
perturbation will be tested for signficance.
}
