% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_exact.R
\name{fisher_exact_thresholded}
\alias{fisher_exact_thresholded}
\title{Fisher exact (thresholded)}
\usage{
fisher_exact_thresholded(
  response_odm,
  grna_odm,
  response_grna_group_pairs,
  control_groups = c("ntc", "compliment"),
  progress = TRUE,
  grna_thresholds = c(1, 3, 5, 7, 8)
)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{grna_odm}{an ODM of either grna expressions (counts) or grna
assignments (logical)}

\item{response_grna_group_pairs}{a data frame with columns \code{response_id} and
\code{grna_group} giving the response ID / grna group pairs to analyze.}

\item{progress}{print progress statements?}
}
\value{
a data frame with columns \code{response_id}, \code{grna_group}, and \code{p_value}.
}
\description{
Carries out Fisher's exact test. \code{grna_odm} is assumed to be a gRNA expression matrix. \code{grna_thresholds}, meanwhile, is an integer vector of gRNA thresholds.
For each gRNA threshold (among those in \code{grna_thresholds}), assigns gRNAs to cells by thresholding the gRNA expression matrix at that level. Then, carries out
a fisher exact test. Cells that cannot be confidently assigned to a given gRNA are excluded.
}
\details{
The \code{grna_group} column of \code{response_grna_group_pairs} contains the names of grna groups.
These grna groups are assumed to be present in the \code{target} column of the feature
covariate matrix of \code{grna_odm}. This column should contain entries "non-targeting"
indicating the non-targeting grnas. If \code{target_type} is present as a column of the
feature covariate matrix of \code{grna_odm}, \code{target_type} is ignored.
}
\examples{
\dontrun{
response_odm <- load_dataset_modality("schraivogel/ground_truth_perturbseq/gene")
grna_odm <- load_dataset_modality("schraivogel/ground_truth_perturbseq/grna_expression")
response_grna_group_pairs <-
 expand.grid(grna_group = c("CCNE2", "CPQ"),
             response_id = sample(ondisc::get_feature_ids(response_odm), 5))
res <- fisher_exact_thresholded(response_odm, grna_odm, response_grna_group_pairs)
}
}
