% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{generate_all_pairs}
\alias{generate_all_pairs}
\title{Generate all pairs}
\usage{
generate_all_pairs(response_odm, grna_odm)
}
\arguments{
\item{response_odm}{a response ODM}

\item{grna_odm}{a gRNA ODM}
}
\value{
a data frame with columns "response_id" and "grna_group" containing all pairs
}
\description{
Generates the entire set of response-gRNA group pairs from an input response ODM and gRNA ODM.
}
