% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{get_control_group_for_method}
\alias{get_control_group_for_method}
\title{Get control group for method}
\usage{
get_control_group_for_method(method)
}
\arguments{
\item{method}{the name of a method}
}
\value{
the control group as a character string, one of "non_targeting_cells" or "all_other_cells"
}
\description{
A given method has a "treatment group" (i.e., the set of cells receiving the perturbation in question)
and a "control group" (i.e., the set of cells against which the treatment group is compared). For some
methods the "control group" is the set of negative control gRNAs, and for other methods the control group
is set of all cells that did not receive the perturbation in question. This function returns the control group
of a given method, either "non_targeting_cells" or "all_other_cells".
}
