% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{get_data_for_pert_prop}
\alias{get_data_for_pert_prop}
\title{Get dataset for perturbation propensity analysis}
\usage{
get_data_for_pert_prop(paper, dataset)
}
\arguments{
\item{paper}{"frangieh", "papalexi", etc.}

\item{dataset}{"ifn_gamma", "eccite_screen", etc.}
}
\value{
Data frame that contains the gRNA assignments and technical covariates.
}
\description{
Get dataset for perturbation propensity analysis.
}
