% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{get_data_method_ram_matrix_from_small_result}
\alias{get_data_method_ram_matrix_from_small_result}
\title{Get data method RAM matrix from small result}
\usage{
get_data_method_ram_matrix_from_small_result(res, p_increase = 1.08)
}
\arguments{
\item{res}{the result data frame}

\item{p_increase}{multiplicative factor by which to increase RAM}
}
\value{
NULL (prints matrix to console)
}
\description{
Get data method RAM matrix from small result
}
