% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{get_grna_assignments_via_max_op}
\alias{get_grna_assignments_via_max_op}
\title{Get grna assignments via max operation}
\usage{
get_grna_assignments_via_max_op(grna_odm)
}
\arguments{
\item{grna_odm}{a grna ODM;}
}
\value{
a vector of grna IDs obtained by applying a column-wise max operation to the grna ODM.
}
\description{
Get grna assignments via max operation
}
