% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{get_gRNA_dataset_name}
\alias{get_gRNA_dataset_name}
\title{Get the name of a gRNA dataset}
\usage{
get_gRNA_dataset_name(dataset_name, gRNA_modality)
}
\arguments{
\item{dataset_name}{the name of a dataset in paper/experiment/modality format}

\item{gRNA_modality}{the name of the gRNA modality, one of "assignment" or "expression"}
}
\value{
the name of the gRNA dataset in paper/experiment/modality format
}
\description{
Get the name of a gRNA dataset
}
