% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{get_library_sizes}
\alias{get_library_sizes}
\title{Get library sizes}
\usage{
get_library_sizes(response_odm)
}
\arguments{
\item{response_odm}{a response ODM}
}
\value{
a vector of cell library sizes
}
\description{
Gets the cell-wise library sizes from a response ODM (either "n_umis" or "n_fragments", as appropriate)
}
