% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{get_sceptre_function_args_for_pair}
\alias{get_sceptre_function_args_for_pair}
\title{Get sceptre function args for pair}
\usage{
get_sceptre_function_args_for_pair(
  response_id,
  undercover_grna,
  dataset_name,
  output_amount,
  B
)
}
\arguments{
\item{response_id}{ID of the response}

\item{undercover_grna}{ID of the gRNA to go undercover}

\item{dataset_name}{name of the dataset}

\item{output_amount}{sceptre output amount}

\item{B}{number of resamples B}
}
\value{
a list of (swapped) arguments to pass (via do.call) to sceptre;
}
\description{
For a given response ID and undercover gRNA, obtains the arguments passed (by do.call) to sceptre in the undercover pipeline.
}
