% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimosca.R
\name{get_sparse_matrix_pieces}
\alias{get_sparse_matrix_pieces}
\title{Get sparse matrix pieces}
\usage{
get_sparse_matrix_pieces(csc_mat)
}
\arguments{
\item{csc_mat}{a sparse matrix in CSC format}
}
\value{
a list containing x, i, p, Dim(1), and Dim(2) (in that order)
}
\description{
Get sparse matrix pieces
}
