% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{get_target_assignments_via_max_op}
\alias{get_target_assignments_via_max_op}
\title{Get target assignments via max operation}
\usage{
get_target_assignments_via_max_op(grna_odm)
}
\arguments{
\item{grna_odm}{a grna ODM}
}
\value{
a vector of "targets" obtained by applying a column-wise max operation to the grna ODM;
the target is present as a column of the feature covariate matrix of the grna ODM.
}
\description{
Get target assignments via max operation
}
\examples{
\dontrun{
grna_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/grna_assignment")
get_target_assignments_via_max_op(grna_odm)
}
}
