% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{get_undercover_groups}
\alias{get_undercover_groups}
\title{Get undercover groups}
\usage{
get_undercover_groups(ntc_names, group_size, partition_count)
}
\arguments{
\item{ntc_names}{a vector of NTC names}

\item{group_size}{size of the undercover groups}

\item{partition_count}{number of undercover groups to generate}
}
\value{
a vector of comma-separated undercover groups
}
\description{
Produce a list of undercover groups given a list of NTCs.
}
