% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liscovitch_method.R
\name{liscovitch_method}
\alias{liscovitch_method}
\title{Liscovitch method}
\usage{
liscovitch_method(response_odm, gRNA_odm, response_gRNA_group_pairs)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{gRNA_odm}{an ODM of gRNA expressions/indicators}

\item{response_gRNA_group_pairs}{a data frame with columns \code{response_id} and \code{gRNA_group} giving the response ID / gRNA group pairs to analyze}
}
\value{
a data frame with columns \code{response_id}, \code{gRNA_group}, and \code{p_value}.
}
\description{
Implements the differential expression method of Liscovitch
}
\examples{

}
