% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{load_dataset_modality}
\alias{load_dataset_modality}
\title{Load dataset modality}
\usage{
load_dataset_modality(
  data_fp,
  offsite_dir = .get_config_path("LOCAL_SCEPTRE2_DATA_DIR")
)
}
\arguments{
\item{data_fp}{(relative) file path to dataset modality}

\item{offsite_dir}{the SCETPRE2 offsite directory}
}
\value{
the (QC'ed) ODM
}
\description{
Load dataset modality
}
\examples{
\dontrun{
schraivogel_enhancer_screen_chr11_gene <-
load_dataset_modality("schraivogel/ground_truth_tapseq/gene")
schraivogel_enhancer_screen_chr11_grna <-
load_dataset_modality("schraivogel/ground_truth_tapseq/grna")
}
}
