% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{load_dataset_multimodal}
\alias{load_dataset_multimodal}
\title{Load dataset multimodal}
\usage{
load_dataset_multimodal(
  paper_fp,
  offsite_dir = .get_config_path("LOCAL_SCEPTRE2_DATA_DIR")
)
}
\arguments{
\item{paper_fp}{(relative) file path to paper modality}

\item{offsite_dir}{the SCETPRE2 offsite directory}
}
\value{
the (QC'ed) multimodal ODM
}
\description{
Load dataset multimodal
}
\examples{
\dontrun{
schraivogel_enhancer_screen_chr11 <-
load_dataset_multimodal("schraivogel/enhancer_screen_chr11")
}
}
