% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{load_whole_odm}
\alias{load_whole_odm}
\title{Load whole odm}
\usage{
load_whole_odm(odm, csc_format = TRUE)
}
\arguments{
\item{odm}{an ondisc_matrix object}

\item{csc_format}{load in cell-accessible, CSC format (TRUE) or feature-accessible, CSR format (FALSE)?}
}
\value{
an in-memory matrix
}
\description{
Loads data from disk into memory by cell.
}
