% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_regression.R
\name{nb_regression}
\alias{nb_regression}
\title{NB regression}
\usage{
nb_regression(
  response_odm,
  grna_odm,
  response_grna_group_pairs,
  progress = TRUE
)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{grna_odm}{an ODM of either grna expressions (counts) or grna
assignments (logical)}

\item{response_grna_group_pairs}{a data frame with columns \code{response_id} and
\code{grna_group} giving the response ID / grna group pairs to analyze.}

\item{progress}{print progress messages?}
}
\value{
a data frame with columns \code{response_id}, \code{grna_group}, and \code{p_value}.
}
\description{
Implements a negative binomial regression.
}
\details{
The \code{grna_group} column of \code{response_grna_group_pairs} contains the names of grna groups.
These grna groups are assumed to be present in the \code{target} column of the feature
covariate matrix of \code{grna_odm}. This column should contain entries "non-targeting"
indicating the non-targeting grnas. If \code{target_type} is present as a column of the
feature covariate matrix of \code{grna_odm}, \code{target_type} is ignored.
}
\examples{
\dontrun{
# a small example
response_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/gene")
grna_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/grna_assignment")
response_grna_group_pairs <-
 expand.grid(grna_group = c("CCNE2", "HS2"),
             response_id = sample(ondisc::get_feature_ids(response_odm), 50))

# a larger example
response_odm <- load_dataset_modality("frangieh/control/gene")
grna_odm <- load_dataset_modality("frangieh/control/grna_assignment")
}
}
