% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schraivogel_helpers.R
\name{normalize_cens_mean}
\alias{normalize_cens_mean}
\title{Censored mean normalization}
\usage{
normalize_cens_mean(
  dge,
  percentile = 0.9,
  norm_counts = FALSE,
  scale_fun = log1p
)
}
\arguments{
\item{dge}{Matrix or data.frame containing digital gene expression data for all cells and genes
to be tested. Rows are genes and columns are cells, with row names being gene ids and column
names cell ids.}

\item{percentile}{Genes of each cell with expression from that percentile upwards will be
excluded to calculate size factors.}

\item{norm_counts}{(logical) Return normalized counts, meaning normalized data is rounded to the
nearest integer to resemble a raw count matrix required by certain DE methods such as DEsingle.
Default = FALSE.}

\item{scale_fun}{Scaling function to be applied to censored mean normalized data. Ignored if
norm_counts = TRUE.}
}
\description{
Normalize DGE data based on censored mean, which excludes genes highly expressed genes when
calculating size factors. This code
was drawn from https://github.com/argschwind/TAPseq_manuscript/blob/master/vignettes/functions/MAST.R
and only lightly edited.
}
