% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weissman.R
\name{odm_to_cell_pop}
\alias{odm_to_cell_pop}
\title{Convert from ODM to CellPopulation format}
\usage{
odm_to_cell_pop(response_odm, pert_assignments, guide_to_target_map)
}
\arguments{
\item{response_odm}{ODM for response variable}

\item{pert_assignments}{Character vector of perturbation assignments, one per cell}
}
\value{
An object of type \code{CellPopulation} that the \code{perturbseq} library expects
}
\description{
Convert from ODM to CellPopulation format
}
