% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{process_multimodal_odm}
\alias{process_multimodal_odm}
\title{Process multimodal odm}
\usage{
process_multimodal_odm(mm_odm)
}
\arguments{
\item{mm_odm}{a multimodal odm}
}
\value{
a streamlined, simpler multimodal odm
}
\description{
Process multimodal odm
}
