% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{process_undercover_result}
\alias{process_undercover_result}
\title{Process undercover results}
\usage{
process_undercover_result(undercover_res, sample_size_df)
}
\arguments{
\item{undercover_res}{an output of the undercover pipeline}

\item{sample_size_df}{the sample sizes data frame}
}
\value{
a proccessed \code{undercover_res}
}
\description{
Processes undercover results. This function (i) replashes slashes with underscores in the dataset names, (ii) combines the Schraivogel enhancer screens, and (iii) appends n treatment (i.e., number of nonzero cells containing the undercover gRNA), n control (i.e., number of nonzero cells containing an NT gRNA, excluding the undercover gRNA), and effective_sample_size (i.e., total number of nonzero cells containing an NT gRNA)
}
