% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{read_all_modalities}
\alias{read_all_modalities}
\title{Read all modalities}
\usage{
read_all_modalities(
  paper,
  dataset,
  sceptre2_data_dir = paste0(.get_config_path("LOCAL_SCEPTRE2_DATA_DIR"), "data/")
)
}
\arguments{
\item{paper}{name of the paper to load}

\item{dataset}{name of the dataset within the paper to load}

\item{sceptre2_data_dir}{location of the sceptre2 data directory}
}
\value{
a multimodal odm for the specified paper/dataset
}
\description{
Read all modalities
}
