% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimosca.R
\name{return_dense_mat}
\alias{return_dense_mat}
\title{Return dense matrix}
\usage{
return_dense_mat(response_odm)
}
\arguments{
\item{response_odm}{an ondisc object}
}
\description{
Returns a dense matrix representation of response_odm
}
\examples{
\dontrun{
response_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/gene")
dense_mat <- return_dense_mat(response_odm)
}
}
