% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{save_all_modalities}
\alias{save_all_modalities}
\title{Save all modalities}
\usage{
save_all_modalities(
  multimodal_odm,
  paper,
  dataset,
  metadata_file_name,
  sceptre2_data_dir = paste0(.get_config_path("LOCAL_SCEPTRE2_DATA_DIR"), "data/")
)
}
\arguments{
\item{multimodal_odm}{a multimodal ODM object}

\item{paper}{name of the paper}

\item{dataset}{name of the dataset within the paper}

\item{metadata_file_name}{name of the metadata file to write (e.g., "metadata_qc.rds")}

\item{sceptre2_data_dir}{location of the sceptre2 data directory}
}
\description{
Save all modalities
}
