% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceptre_functs.R
\name{sceptre_no_covariates}
\alias{sceptre_no_covariates}
\title{SCEPTRE (no covariates)}
\usage{
sceptre_no_covariates(
  response_odm,
  grna_odm,
  response_grna_group_pairs,
  B = 2500,
  output_amount = 1,
  distilled = FALSE
)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{grna_odm}{an ODM of either grna expressions (counts) or grna
assignments (logical)}

\item{response_grna_group_pairs}{a data frame with columns \code{response_id} and
\code{grna_group} giving the response ID / grna group pairs to analyze.}

\item{B}{number of resamples to draw}

\item{output_amount}{amount of output to return from function (ranging from 1-3)}

\item{distilled}{use the distilled statistic (TRUE) or the full statistic (FALSE)?}
}
\description{
Runs SCEPTRE w/o covariates
}
