% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schraivogel_method.R
\name{schraivogel_method}
\alias{schraivogel_method}
\title{Run Schraivogel's MAST.cov method}
\usage{
schraivogel_method(
  gene_odm_fp,
  gene_metadata_fp,
  grna_odm_fp,
  grna_metadata_fp,
  gene_gRNA_group_pairs,
  gRNA_groups_table = NULL,
  gRNA_threshold = 8
)
}
\arguments{
\item{gene_odm_fp}{Filepath for gene expression ODM}

\item{gene_metadata_fp}{Filepath for gene expression ODM metadata}

\item{grna_odm_fp}{Filepath for gRNA expression / indicator ODM}

\item{grna_metadata_fp}{Filepath for gRNA expression / indicator ODM metadata}

\item{gene_gRNA_group_pairs}{Pairs of genes and gRNA groups to analyze, as in \code{sceptre}}

\item{gRNA_groups_table}{A table specifying which gRNAs are in which groups, as in \code{sceptre}.
This argument is optional, and the default assumption is that each gRNA is in its own group.}

\item{gRNA_threshold}{A threshold for gRNA expression. This argument is optional, and defaults to 8,
which was Schraivogel et al's choice.}
}
\value{
A tibble with three columns: \code{gene_id}, \code{gRNA_group}, \code{p_value}.
The last column is the computed MAST.cov p-value.
}
\description{
Run Schraivogel's MAST.cov method
}
