% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowmoi.R
\name{abstract_interface}
\alias{abstract_interface}
\title{Abstract method interface}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{gRNA_odm}{an ODM of gRNA expressions/indicators}

\item{response_gRNA_group_pairs}{a data frame with columns \code{response_id} and \code{gRNA_group} giving the response ID / gRNA group pairs to analyze}
}
\value{
a data frame with columns \code{response_id}, \code{gRNA_group}, and \code{p_value}.
}
\description{
Abstract method interface
}
\examples{
\dontrun{
gene_odm <- load_dataset_modality("frangieh/co_culture/gene")
gRNA_odm <- load_dataset_modality("frangieh/co_culture/grna")
response_gRNA_group_pairs <-
data.frame(response_id = sample(ondisc::get_feature_ids(gene_odm), 500, FALSE),
gRNA_group = sample(ondisc::get_feature_ids(gRNA_odm), 500, FALSE))
}
}
