% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_methods.R
\name{dummy_method_1}
\alias{dummy_method_1}
\title{Dummy method 1}
\usage{
dummy_method_1(
  response_odm,
  gRNA_odm,
  response_gRNA_group_pairs,
  dist = "runif"
)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{gRNA_odm}{an ODM of either gRNA expressions (counts) or gRNA
assignments (logical)}

\item{response_gRNA_group_pairs}{a data frame with columns \code{response_id} and
\code{gRNA_group} giving the response ID / gRNA group pairs to analyze.}

\item{dist}{a character vector giving the distribution to generate random variates from}
}
\value{
a data frame with columns \code{response_id}, \code{gRNA_group}, and \code{p_value}.
}
\description{
A dummy method; samples uniformly distributed p-values
}
\details{
The \code{gRNA_group} column of \code{response_gRNA_group_pairs} contains the names of gRNA groups.
These gRNA groups are assumed to be present in the \code{target} column of the feature
covariate matrix of \code{gRNA_odm}. This column should contain entries "non-targeting"
indicating the non-targeting gRNAs. If \code{target_type} is present as a column of the
feature covariate matrix of \code{gRNA_odm}, \code{target_type} is ignored.
}
\examples{
\dontrun{
response_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/gene")
gRNA_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/grna_assignment")
response_gRNA_group_pairs <-
 expand.grid(gRNA_group = c("CCNE2-TSS", "HS2-enh"),
             response_id = sample(ondisc::get_feature_ids(response_odm), 50))
}
}
