% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_methods.R
\name{dummy_method_1}
\alias{dummy_method_1}
\title{Dummy method 1}
\usage{
dummy_method_1(response_odm, gRNA_odm, response_gRNA_group_pairs)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{gRNA_odm}{an ODM of gRNA expressions/indicators}

\item{response_gRNA_group_pairs}{a data frame with columns \code{response_id} and \code{gRNA_group} giving the response ID / gRNA group pairs to analyze}
}
\value{

}
\description{
A dummy method; samples uniformly distributed p-values
}
\examples{

}
