% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{get_gRNA_assignments_via_max_op}
\alias{get_gRNA_assignments_via_max_op}
\title{Get gRNA assignments via max operation}
\usage{
get_gRNA_assignments_via_max_op(gRNA_odm)
}
\arguments{
\item{gRNA_odm}{a gRNA ODM;}
}
\value{
a vector of gRNA IDs obtained by applying a column-wise max operation to the gRNA ODM.
}
\description{
Get gRNA assignments via max operation
}
