% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functs.R
\name{get_grna_dataset_name}
\alias{get_grna_dataset_name}
\title{Get the name of a grna dataset}
\usage{
get_grna_dataset_name(dataset_name, grna_modality)
}
\arguments{
\item{dataset_name}{the name of a dataset in paper/experiment/modality format}

\item{grna_modality}{the name of the grna modality, one of "assignment" or "expression"}
}
\value{
the name of the grna dataset in paper/experiment/modality format
}
\description{
Get the name of a grna dataset
}
