% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liscovitch_method.R
\name{liscovitch_method}
\alias{liscovitch_method}
\title{Liscovitch method}
\usage{
liscovitch_method(response_odm, gRNA_odm, response_gRNA_group_pairs)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{gRNA_odm}{an ODM of either gRNA expressions (counts) or gRNA
assignments (logical)}

\item{response_gRNA_group_pairs}{a data frame with columns \code{response_id} and
\code{gRNA_group} giving the response ID / gRNA group pairs to analyze.}
}
\value{
a data frame with columns \code{response_id}, \code{gRNA_group}, and \code{p_value}.
}
\description{
Implements the differential expression method of Liscovitch
}
\details{
The \code{gRNA_group} column of \code{response_gRNA_group_pairs} contains the names of gRNA groups.
These gRNA groups are assumed to be present in the \code{target} column of the feature
covariate matrix of \code{gRNA_odm}. This column should contain entries "non-targeting"
indicating the non-targeting gRNAs. If \code{target_type} is present as a column of the
feature covariate matrix of \code{gRNA_odm}, \code{target_type} is ignored.
}
\examples{
\dontrun{
response_odm <- load_dataset_modality("liscovitch/experiment_small/chromatin")
gRNA_odm <- load_dataset_modality("liscovitch/experiment_small/grna")
response_gRNA_group_pairs <- expand.grid(response_id = response_odm |> ondisc::get_feature_ids(),
gRNA_group = (gRNA_odm |> ondisc::get_feature_ids())[1:2])
res <- liscovitch_method(response_odm, gRNA_odm, response_gRNA_group_pairs)
}
}
