% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimosca.R
\name{mimosca}
\alias{mimosca}
\title{Mimosca
Implements the MIMOSCA method.}
\usage{
mimosca(response_odm, gRNA_odm, response_gRNA_group_pairs)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{gRNA_odm}{an ODM of gRNA expressions/indicators}

\item{response_gRNA_group_pairs}{a data frame with columns \code{response_id} and \code{gRNA_group} giving the response ID / gRNA group pairs to analyze}
}
\value{
a data frame with columns \code{response_id}, \code{gRNA_group}, and \code{p_value}.
}
\description{
Mimosca
Implements the MIMOSCA method.
}
\examples{
response_odm <- load_dataset_modality("frangieh/control/gene")
response_odm <- response_odm[1:100,]
gRNA_odm <- load_dataset_modality("frangieh/control/grna")
response_gRNA_group_pairs <- data.frame(response_id = response_odm |> get_feature_ids(),
gRNA_group = "TTGCGGCCTCGATACGATAT")
}
