% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimosca.R
\name{mimosca}
\alias{mimosca}
\title{Mimosca
Implements the MIMOSCA method.}
\usage{
mimosca(response_odm, gRNA_odm, response_gRNA_group_pairs, n_rep = 1000)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{gRNA_odm}{an ODM of either gRNA expressions (counts) or gRNA
assignments (logical)}

\item{response_gRNA_group_pairs}{a data frame with columns \code{response_id} and
\code{gRNA_group} giving the response ID / gRNA group pairs to analyze. NOTE: the
gRNA groups currently are assumed to be singleton groups, specified by their
gRNA IDs. Adding functionality to support more general gRNA groups is on our
longer-term to do list.}

\item{n_rep}{number of permutations to use}
}
\value{
a data frame with columns \code{response_id}, \code{gRNA_group}, and \code{p_value}.
}
\description{
Mimosca
Implements the MIMOSCA method.
}
\examples{
response_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/gene")
gRNA_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/grna_assignment")
response_gRNA_group_pairs <-
 expand.grid(response_id = sample(ondisc::get_feature_ids(response_odm)),
            gRNA_group = sample(ondisc::get_feature_ids(gRNA_odm), 2))
}
