% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weissman.R
\name{odm_to_cell_pop}
\alias{odm_to_cell_pop}
\title{Converse from ODM to CellPopulation format}
\usage{
odm_to_cell_pop(response_odm, gRNA_odm)
}
\arguments{
\item{response_odm}{ODM for response variable}

\item{gRNA_odm}{ODM for gRNAs}
}
\value{
An object of type \code{CellPopulation} that the \code{perturbseq} library expects
}
\description{
Converse from ODM to CellPopulation format
}
