% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{process_pc_result}
\alias{process_pc_result}
\title{Process positive control result}
\usage{
process_pc_result(pc_res, sample_size_df)
}
\arguments{
\item{pc_res}{positive control result data frame}

\item{sample_size_df}{sample size data frame}
}
\value{
a processed \code{pc_res}
}
\description{
Process positive control result
}
