% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schraivogel_method.R
\name{schraivogel_method}
\alias{schraivogel_method}
\title{Run Schraivogel's MAST.cov method}
\usage{
schraivogel_method(
  response_odm,
  gRNA_odm,
  response_gRNA_group_pairs,
  gRNA_groups_table = NULL,
  gRNA_threshold = 8
)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{gRNA_odm}{an ODM of either gRNA expressions (counts) or gRNA
assignments (logical)}

\item{response_gRNA_group_pairs}{a data frame with columns \code{response_id} and
\code{gRNA_group} giving the response ID / gRNA group pairs to analyze.}

\item{gRNA_groups_table}{A table specifying which gRNAs are in which groups, as in \code{sceptre}.
This argument is optional, and the default assumption is that each gRNA is in its own group.}

\item{gRNA_threshold}{A threshold for gRNA expression. This argument is optional, and defaults to 8,
which was Schraivogel et al's choice.}
}
\value{
a data frame with columns \code{response_id}, \code{gRNA_group}, and \code{p_value}.
}
\description{
Runs Schraivogel's MAST.cov method.
}
\details{
The \code{gRNA_group} column of \code{response_gRNA_group_pairs} contains the names of gRNA groups.
These gRNA groups are assumed to be present in the \code{target} column of the feature
covariate matrix of \code{gRNA_odm}. This column should contain entries "non-targeting"
indicating the non-targeting gRNAs. If \code{target_type} is present as a column of the
feature covariate matrix of \code{gRNA_odm}, \code{target_type} is ignored.
}
\examples{
\dontrun{
response_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/gene")
gRNA_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/grna")
response_gRNA_group_pairs <- expand.grid(response_id = (response_odm |> ondisc::get_feature_ids()), gRNA_group = c("GATA1-C", "GATA1-D"))
result <- weissman(response_odm, gRNA_odm, response_gRNA_group_pairs)
}
}
