% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schraivogel_method.R
\name{schraivogel_method}
\alias{schraivogel_method}
\title{Run Schraivogel's MAST.cov method}
\usage{
schraivogel_method(
  response_odm,
  gRNA_odm,
  response_gRNA_group_pairs,
  gRNA_groups_table = NULL,
  gRNA_threshold = 8
)
}
\arguments{
\item{response_odm}{gene (or more generally, response) expression ODM}

\item{gRNA_odm}{gRNA expression / indicator ODM}

\item{response_gRNA_group_pairs}{Pairs of responses and gRNA groups to analyze, as in \code{sceptre}}

\item{gRNA_groups_table}{A table specifying which gRNAs are in which groups, as in \code{sceptre}.
This argument is optional, and the default assumption is that each gRNA is in its own group.}

\item{gRNA_threshold}{A threshold for gRNA expression. This argument is optional, and defaults to 8,
which was Schraivogel et al's choice.}
}
\value{
A tibble with three columns: \code{response_id}, \code{gRNA_group}, \code{p_value}.
The last column is the computed MAST.cov p-value.
}
\description{
Run Schraivogel's MAST.cov method
}
