% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schraivogel_method.R
\name{schraivogel_method}
\alias{schraivogel_method}
\title{Run Schraivogel's MAST.cov method}
\usage{
schraivogel_method(
  response_odm,
  grna_odm,
  response_grna_group_pairs,
  grna_threshold = 8
)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{grna_odm}{an ODM of either grna expressions (counts) or grna
assignments (logical)}

\item{response_grna_group_pairs}{a data frame with columns \code{response_id} and
\code{grna_group} giving the response ID / grna group pairs to analyze.}

\item{grna_threshold}{A threshold for grna expression. This argument is optional and defaults to 8,
which was Schraivogel et al's choice.}
}
\value{
a data frame with columns \code{response_id}, \code{grna_group}, and \code{p_value}.
}
\description{
Runs Schraivogel's MAST.cov method.
}
\details{
The \code{grna_group} column of \code{response_grna_group_pairs} contains the names of grna groups.
These grna groups are assumed to be present in the \code{target} column of the feature
covariate matrix of \code{grna_odm}. This column should contain entries "non-targeting"
indicating the non-targeting grnas. If \code{target_type} is present as a column of the
feature covariate matrix of \code{grna_odm}, \code{target_type} is ignored.
}
\examples{
\dontrun{
# a small example
response_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/gene")
grna_odm <- load_dataset_modality("schraivogel/ground_truth_tapseq/grna_assignment")
response_grna_group_pairs <-
 expand.grid(grna_group = c("CCNE2", "HS2"),
             response_id = sample(ondisc::get_feature_ids(response_odm), 50))

# a larger example
response_odm <- load_dataset_modality("frangieh/control/gene")
grna_odm <- load_dataset_modality("frangieh/control/grna_assignment")
response_grna_group_pairs <- data.frame(grna_group = "A2M",
                                        response_id = sample(ondisc::get_feature_ids(response_odm), 1))
}
}
