% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_de.R
\name{seurat_de}
\alias{seurat_de}
\title{Seurat DE}
\usage{
seurat_de(response_odm, gRNA_odm, response_gRNA_group_pairs)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{gRNA_odm}{an ODM of gRNA expressions/indicators}

\item{response_gRNA_group_pairs}{a data frame with columns \code{response_id} and \code{gRNA_group} giving the response ID / gRNA group pairs to analyze}
}
\value{
a data frame with columns \code{response_id}, \code{gRNA_group}, and \code{p_value}.
}
\description{
Implements the Suerat differential expression method.
}
