% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_de.R
\name{seurat_de}
\alias{seurat_de}
\title{Seurat DE}
\usage{
seurat_de(response_odm, gRNA_odm, response_gRNA_group_pairs)
}
\arguments{
\item{response_odm}{an expression ODM of responses (typically genes)}

\item{gRNA_odm}{an ODM of either gRNA expressions (counts) or gRNA
assignments (logical)}

\item{response_gRNA_group_pairs}{a data frame with columns \code{response_id} and
\code{gRNA_group} giving the response ID / gRNA group pairs to analyze. NOTE: the
gRNA groups currently are assumed to be singleton groups, specified by their
gRNA IDs. Adding functionality to support more general gRNA groups is on our
longer-term to do list.}
}
\value{
a data frame with columns \code{response_id}, \code{gRNA_group}, and \code{p_value}.
}
\description{
Implements the Seurat differential expression method.
}
\examples{
\dontrun{
response_odm <- load_dataset_modality("frangieh/co_culture/gene")
gRNA_odm <- load_dataset_modality("frangieh/co_culture/grna_assignment")
response_gRNA_group_pairs <-
data.frame(response_id = sample(ondisc::get_feature_ids(gene_odm), 500, FALSE),
gRNA_group = sample(ondisc::get_feature_ids(gRNA_odm), 500, FALSE))
}
}
